ci: move util function to common utils file
