+++ packages/stg-evm-v2/devtools/config/mainnet/01/credit-messaging.config.ts
@@ -2,8 +2,8 @@ import { CreditMessagingEdgeConfig, CreditMessagingNodeConfig } from '@stargatef
 
 import { OmniGraphHardhat, createGetHreByEid } from '@layerzerolabs/devtools-evm-hardhat'
 
-import { generateCreditMessagingConfig, getSafeAddress } from '../../utils'
-import { filterConnections, getContracts, isValidCreditMessagingChain, validCreditMessagingChains } from '../utils'
+import { filterConnections, generateCreditMessagingConfig, getSafeAddress } from '../../utils'
+import { getContracts, isValidCreditMessagingChain, validCreditMessagingChains } from '../utils'
 
 import { DEFAULT_PLANNER } from './constants'
 import { getMessagingAssetConfig } from './shared'
